projects/identity_access_management/backend/requirements.txt
astroid==3.0.0
autopep8==2.0.4
blinker==1.6.2
click==8.1.7
dill==0.3.7
ecdsa==0.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
future==0.18.3
greenlet==3.0.0
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
pip==23.2.1
platformdirs==3.11.0
pyasn1==0.5.0
pycodestyle==2.11.0
pycrypto==2.6.1
pylint==3.0.1
python-jose==3.3.0
python-jose-cryptodome==1.3.2
rsa==4.9
setuptools==68.0.0
six==1.16.0
SQLAlchemy==2.0.21
tomli==2.0.1
tomlkit==0.12.1
typed-ast==1.5.5
typing_extensions==4.8.0
Werkzeug==3.0.0
wheel==0.41.2
wrapt==1.15.0
zipp==3.17.0
